Created a separate redux store for user's data
