+++ src/store/actions/actionTypes.js
@@ -2,9 +2,10 @@ export const AUTH_START = 'AUTH_START';
 export const AUTH_SUCCESS = 'AUTH_SUCCESS';
 export const AUTH_FAIL = 'AUTH_FAIL';
 export const AUTH_LOGOUT = 'AUTH_LOGOUT';
-export const SET_ONLINE_USER = 'SET_ONLINE_USER';
 
 export const SET_AUTH_REDIRECT_PATH = 'SET_AUTH_REDIRECT_PATH';
 
 export const SET_GROUPS = 'SET_GROUPS';
 export const SET_DEFAULT_GROUP = 'SET_DEFAULT_GROUP';
+
+export const SET_ONLINE_USER = 'SET_ONLINE_USER';
