+++ src/store/actions/auth.js
@@ -15,15 +15,6 @@ export const authSuccess = (token, userId) => {
     };
 };
 
-export const setOnlineUser = (username, email, photo) => {
-    return {
-        type: actionTypes.SET_ONLINE_USER,
-        username,
-        email,
-        photo
-    }
-}
-
 export const authFail = (error) => {
     return {
         type: actionTypes.AUTH_FAIL,
@@ -80,16 +71,6 @@ export const auth = (user, isSignup) => {
             });
     };
 };
-
-export const fetchUserData = (userId) => {
-    return async dispatch => {
-        return await api.get(`/users/${userId}`)
-        .then(response => {      
-            dispatch(setOnlineUser(response.data.user.username, response.data.user.email, response.data.photo));
-        })
-        .catch(error => {console.log(error)});
-    };
-}
 
 export const authCheckState = () => {
     return dispatch => {
