+++ src/store/reducers/auth.js
@@ -4,9 +4,6 @@ import { updateObject } from '../utility';
 const initialState = {
     token: null,
     userId: null,
-    userName: null,
-    email: null,
-    photo: null,
     error: null,
     loading: false,
     authRedirectPath: '/'
@@ -22,14 +19,6 @@ const authSuccess = (state, action) => {
         userId: action.userId,
         error: null,
         loading: false
-    });
-}
-
-const setOnlineUser = (state, action) => {
-    return updateObject(state, {
-        username: action.username,
-        email: action.email,
-        photo: action.photo
     });
 }
 
@@ -60,8 +49,6 @@ const reducer = (state = initialState, action) => {
             return authLogout(state, action);
         case actionTypes.SET_AUTH_REDIRECT_PATH:
             return setAuthRedirectPath(state, action);
-        case actionTypes.SET_ONLINE_USER:
-            return setOnlineUser(state, action);
         default:
             return state;
     }
