+++ src/store/reducers/user.js
@@ -1,0 +1,29 @@
+import * as actionTypes from '../actions/actionTypes';
+import { updateObject } from '../utility';
+
+const initialState = {
+    userId: null,
+    userName: null,
+    email: null,
+    photo: null
+};
+
+const setOnlineUser = (state, action) => {
+  return updateObject(state, {
+      userId: action.id,
+      username: action.username,
+      email: action.email,
+      photo: action.photo
+  });
+}
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+      case actionTypes.SET_ONLINE_USER:
+          return setOnlineUser(state, action);
+      default:
+          return state;
+  }
+};
+
+export default reducer;
